+++ changelog.md
@@ -1,7 +1,7 @@
 # Changelog for NovaSheets
 - [1.0.0](#100)
   - [1.0.0-pre1](#100-pre1) • [1.0.0-pre2](#100-pre2) • [1.0.0-pre3](#100-pre3) • [1.0.0-pre4](#100-pre4)
-  • [1.0.0-rc1](#100-rc1) • [1.0.0-rc2](#100-rc2) • [1.0.0-rc3](#100-rc3)
+  • [1.0.0-rc1](#100-rc1) • [1.0.0-rc2](#100-rc2) • [1.0.0-rc3](#100-rc3) • [1.0.0-rc4](#100-rc4)
 - [0.6.0](#060) • [0.6.1](#061) • [0.6.2](#062) • [0.6.3](#063) • [0.6.4](#064) • [0.6.5](#065) • [0.6.6](#066) • [0.6.7](#067)
 - [0.5.0](#050) • [0.5.1](#051) • [0.5.2](#052)
 - [0.4.0](#040) • [0.4.1](#041) • [0.4.2](#042) • [0.4.3](#043) • [0.4.4](#044) • [0.4.5](#045) • [0.4.6](#046) • [0.4.7](#047)
@@ -11,6 +11,16 @@
 
 ## 1.0.0
 *Upcoming*
+
+### 1.0.0-rc4
+*2021-05-29*
+- **Changes**
+  - [Syntax] Changed simple breakpoints to not output `only screen`.
+- **Fixes**
+  - [Syntax] Fixed vanilla CSS functions not being parsed when containing NovaSheets content.
+  - [Syntax] Fixed CSS `calc()` functions not being parsed properly.
+  - [Syntax] Fixed comments being duplicated over child selectors.
+  - [Functions] Fixed `@breakpoint` function not being able to be used in a child selector by making it use simple breakpoints.
 
 ### 1.0.0-rc3
 *2021-05-22*